add calendar sidebar

+++ todo-angular/frontend/src/app/shared/edit-todo/edit-todo.component.ts
@@ -1,8 +1,9 @@
-import {Component, Input} from '@angular/core';
+import { Component, Input } from '@angular/core';
 import { TodosService } from '../../services/todos.service';
 import { Dialog } from '@angular/cdk/dialog';
 import { ModalWindowComponent } from '../modal-window/modal-window.component';
 import { ModalListNames } from '../../interfaces';
+import { ItemsService } from '../../services/items.service';
 
 @Component({
   selector: 'app-edit-todo',
@@ -10,16 +11,22 @@ import { ModalListNames } from '../../interfaces';
   styleUrls: ['./edit-todo.component.scss'],
 })
 export class EditTodoComponent {
-  @Input('todoTitle') todoTitle?:string;
-  @Input('todoId') todoId?:string;
+  @Input('todoTitle') todoTitle?: string;
+  @Input('id') id?: string;
+  @Input('type') type?: string;
   readonly modalListNames = ModalListNames;
-  constructor(private todoService: TodosService, private dialog: Dialog) {}
+  constructor(private todoService: TodosService, private dialog: Dialog, private itemService: ItemsService) {}
 
   deleteTodo(): void {
-    this.todoService.deleteTodo();
+    if (this.type === 'todo') this.todoService.deleteTodo();
+    if (this.type === 'item') this.itemService.deleteItem(this.id);
   }
-  openModalWindow(windowName: string, value?: string): void {
-    const dialogRef = this.dialog.open(ModalWindowComponent, { data: { name: windowName, id:this.todoId } });
+  openModalWindow(): void {
+    const dialogRef =
+      this.type === 'todo'
+        ? this.dialog.open(ModalWindowComponent, { data: { name: this.modalListNames.editTodo, id: this.id } })
+        : this.dialog.open(ModalWindowComponent, { data: { name: this.modalListNames.editItem, id: this.id } });
+
     dialogRef.closed.subscribe();
   }
 }
